fix: change auth logic
