+++ src/main.tsx
@@ -4,7 +4,7 @@ import { Provider } from '@/components/ui/provider';
 import './index.css';
 import { RouterProvider } from 'react-router';
 import { router } from './router/routes';
-import { AuthProvider } from './store/session-provider';
+import { AuthProvider } from './store/auth-provider';
 
 const rootElement = document.querySelector('#root');
